refactor: rename Contributed* to simpler names
BREAKING CHANGE: Module paths changed.
